// packages/babylon-vrm-loader/src/VRM/vrm-interfaces.ts
export interface IVRMVector3 {
    x: number;
    y: number;
    z: number;
}
// ...
export interface IVRMSecondaryAnimationSpring {
    comment: string;
    stiffiness: number;
    gravityPower: number;
    gravityDir: IVRMVector3;
    dragForce: number;
    center: number;
    hitRadius: number;
    bones: number[];
    colliderGroups: number[];
}

export interface IVRMSecondaryAnimationCollider {
    offset: IVRMVector3;
    radius: number;
}

export interface IVRMSecondaryAnimationColliderGroup {
    node: number;
    colliders: IVRMSecondaryAnimationCollider[];
}

/**
 * extensions.VRM.secondaryAnimation
 */
export interface IVRMSecondaryAnimation {
    boneGroups: IVRMSecondaryAnimationSpring[];
    colliderGroups: IVRMSecondaryAnimationColliderGroup[];
}